feat: finish compiler for only constant
For constant, we will produce the byte code `op index64_high index64_low`.
Do remember we do not produce the constant itself, but produce the
index. And next we make the vm, it is easy, it will use `ip` to traverse
the instructions, and push the constant to the stack.

It is not easy to get here. well...