feat: add binding support
In this commit, the most important thing is to change all the
`unique_ptr` to the `shared_ptr`. It is not convenient to use
`unique_ptr` which would cause much more unnecessary copy. However, this
time I have made mistakes just like in the interpreter.

In order to achieving bindings, we need to first create symbol table, it
is just like the environment in the interpreter (the same idea).
However, in the VM, we should store the pointer. This is really boring
because we are using cpp, we need to handle memory.