feat: add support for basic function compilation
At this commit, we only support the following function compilation:

```monkey
fn() { return 1 + 2; }
```

This commit mainly focuses at writing template tests. It it hard to do
this.

And then we need to add scopes to generate each instruction. It is not
difficult. However, it would be complicated when using bindings (TODO)