feat: add support for local bindings
local variable would also in the stack, so we need to think in what
situations would produce local variables. Yes, when we enter a new
frame. But we need a new symbol table for getting the mapping from name
to the index.

Now the question is the the new symbolTable's index would start from 0,
and we need a way to utilize the original stack. We could utilize the
information that symbolTable has. We can use the `numDefinitions` field
to get how many local variables.

Thus we could reverse the size for the local variables like below:

```
sp here

variable 2
variable 1
OpCall
```

It's a basic idea. This is what the real system done.