feat: add builtin function support for compiler
The code first refactor the built function in the evaluator to the
`builtins.hpp`.

It's not hard to achieve builtins. First, we need to define the symbol
in the global. And next we should create a new symbol scope for builtins
and next we will exactly what we have done.